ml_models/inference.py
"""
ML Model Inference Module for Smart Bus Stop
Provides inference capabilities for image analysis
"""

import os
import numpy as np
from PIL import Image
import torch
import torchvision.transforms as transforms
from dotenv import load_dotenv
import logging

# Load environment variables
load_dotenv()

# Setup logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class MLInferenceHandler:
    """Handler for ML model inference operations"""
    
    def __init__(self, model_path=None):
        """
        Initialize ML inference handler
        
        Args:
            model_path (str): Path to the model file
        """
        self.model_path = model_path or os.getenv('MODEL_PATH', './ml_models/')
        self.confidence_threshold = float(os.getenv('CONFIDENCE_THRESHOLD', 0.5))
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        
        # Image preprocessing
        self.transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
        ])
        
        logger.info(f"ML Inference initialized on device: {self.device}")
        
        # Placeholder for model loading
        self.model = None
        # Uncomment and modify when you have a specific model:
        # self.load_model()
    
    def load_model(self, model_name='model.pth'):
        """
        Load ML model from file
        
        Args:
            model_name (str): Name of the model file
        """
        try:
            model_file = os.path.join(self.model_path, model_name)
            
            if not os.path.exists(model_file):
                logger.warning(f"Model file not found: {model_file}")
                return False
            
            # Example: Load a PyTorch model
            # self.model = torch.load(model_file, map_location=self.device)
            # self.model.eval()
            
            logger.info(f"Model loaded: {model_name}")
            return True
        
        except Exception as e:
            logger.error(f"Error loading model: {str(e)}")
            return False
    
    def preprocess_image(self, image_path):
        """
        Preprocess image for inference
        
        Args:
            image_path (str): Path to image file
        
        Returns:
            torch.Tensor: Preprocessed image tensor
        """
        try:
            image = Image.open(image_path).convert('RGB')
            image_tensor = self.transform(image).unsqueeze(0)
            return image_tensor.to(self.device)
        
        except Exception as e:
            logger.error(f"Error preprocessing image: {str(e)}")
            raise
    
    def infer(self, image_path):
        """
        Run inference on an image
        
        Args:
            image_path (str): Path to image file
        
        Returns:
            dict: Inference results
        """
        try:
            if self.model is None:
                # Return placeholder results when model is not loaded
                return {
                    'status': 'no_model',
                    'message': 'Model not loaded. Please load a model first.',
                    'predictions': [],
                    'confidence': 0.0
                }
            
            # Preprocess image
            image_tensor = self.preprocess_image(image_path)
            
            # Run inference
            with torch.no_grad():
                outputs = self.model(image_tensor)
                
                # Process outputs (example for classification)
                probabilities = torch.nn.functional.softmax(outputs, dim=1)
                confidence, predicted = torch.max(probabilities, 1)
                
                result = {
                    'status': 'success',
                    'predictions': [
                        {
                            'class_id': predicted.item(),
                            'confidence': confidence.item()
                        }
                    ],
                    'confidence': confidence.item()
                }
            
            logger.info(f"Inference completed for {image_path}")
            return result
        
        except Exception as e:
            logger.error(f"Error during inference: {str(e)}")
            return {
                'status': 'error',
                'message': str(e),
                'predictions': [],
                'confidence': 0.0
            }
    
    def batch_infer(self, image_paths):
        """
        Run inference on multiple images
        
        Args:
            image_paths (list): List of image file paths
        
        Returns:
            list: List of inference results
        """
        results = []
        for image_path in image_paths:
            result = self.infer(image_path)
            results.append(result)
        
        return results
    
    def detect_objects(self, image_path):
        """
        Object detection inference (placeholder)
        
        Args:
            image_path (str): Path to image file
        
        Returns:
            dict: Detection results with bounding boxes
        """
        # Placeholder for object detection
        # This would use models like YOLO, Faster R-CNN, etc.
        return {
            'status': 'not_implemented',
            'message': 'Object detection not yet implemented',
            'detections': []
        }
    
    def analyze_crowd(self, image_path):
        """
        Crowd density analysis (placeholder)
        
        Args:
            image_path (str): Path to image file
        
        Returns:
            dict: Crowd analysis results
        """
        # Placeholder for crowd analysis
        return {
            'status': 'not_implemented',
            'message': 'Crowd analysis not yet implemented',
            'crowd_density': 0,
            'people_count': 0
        }


if __name__ == '__main__':
    # Test inference handler
    handler = MLInferenceHandler()
    logger.info("ML Inference Handler initialized successfully")
    
    # Test with a sample image (if exists)
    test_image = './test_image.jpg'
    if os.path.exists(test_image):
        result = handler.infer(test_image)
        logger.info(f"Test inference result: {result}")
    else:
        logger.info("No test image found. Skipping inference test.")
